Create RefNamePrivate.hpp file
