+++ libGitWrap/RefName.hpp
@@ -19,6 +19,7 @@
 
 #ifndef GIT_REF_NAME_HPP
 #define GIT_REF_NAME_HPP
+#pragma once
 
 #include <QStringList>
 #include <QRegExp>
